added seeds for recipes and dishes

+++ data/migrations/20190606121555_create_receipe-ingredients_table.js
@@ -21,7 +21,7 @@ exports.up = function(knex, Promise) {
                 .onUpdate('CASCADE')
 
             tbl //needed quantity of ingredient
-                .integer('quantity')
+                .float('quantity')
                 .notNullable()
         })
